修复在本地ubuntu虚拟机执行预下载时npm install报错的问题

+++ prebuilts_download.py
@@ -236,43 +236,56 @@ def _npm_install(args):
     out, err, retcode = _run_cmd(npm_package_lock_cmd)
     if retcode != 0:
         return False, err.decode()
-    procs = []
-    full_code_path_list = []
-    cmd_list = []
+    install_cmds = []
+    full_code_paths = []
+            
     print('start npm install, please wait.')
     for install_info in args.npm_install_config:
         full_code_path = os.path.join(args.code_dir, install_info)
+        if full_code_path in args.success_installed:
+            print('{} has been installed, skip'.format(full_code_path))
+            continue
         basename = os.path.basename(full_code_path)
         node_modules_path = os.path.join(full_code_path, "node_modules")
         npm_cache_dir = os.path.join('~/.npm/_cacache', basename)
         if os.path.exists(node_modules_path):
             print('remove node_modules %s' % node_modules_path)
             _run_cmd(('rm -rf {}'.format(node_modules_path)))
-    for install_info in args.npm_install_config:
-        full_code_path = os.path.join(args.code_dir, install_info)
-        basename = os.path.basename(full_code_path)
-        node_modules_path = os.path.join(full_code_path, "node_modules")
-        npm_cache_dir = os.path.join('~/.npm/_cacache', basename)
         if os.path.exists(full_code_path):
             cmd = ['timeout', '-s', '9', '90s', npm, 'install', '--registry', args.npm_registry, '--cache', npm_cache_dir]
             if args.host_platform == 'darwin':
                 cmd = [npm, 'install', '--registry', args.npm_registry, '--cache', npm_cache_dir]
             if args.unsafe_perm:
                 cmd.append('--unsafe-perm')
-            proc = subprocess.Popen(cmd, cwd=full_code_path, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-            # wait proc Popen with 0.1 second
-            time.sleep(0.1)
-            procs.append(proc)
-            full_code_path_list.append(full_code_path)
-            cmd_list.append(cmd)
+            install_cmds.append(cmd)
+            full_code_paths.append(full_code_path)
         else:
             raise Exception("{} not exist, it shouldn't happen, pls check...".format(full_code_path))
-    for ind, proc in enumerate(procs):
-        out, err = proc.communicate()
-        if proc.returncode:
-            print("in dir:{}, executing:{}".format(full_code_path_list[ind], ' '.join(cmd_list[ind])))
-            return False, err.decode()
+    
+    if args.parallel_install:
+        procs = []
+        for index, install_cmd in enumerate(install_cmds):
+            proc = subprocess.Popen(install_cmd, cwd=full_code_paths[index], stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+            print('run npm install in {}'.format(full_code_paths[index]))
+            time.sleep(0.1)
+            procs.append(proc)
+        for index, proc in enumerate(procs):
+            out, err = proc.communicate()
+            if proc.returncode:
+                print("in dir:{}, executing:{}".format(full_code_paths[index], ' '.join(install_cmd[index])))
+                return False, err.decode()
+            args.success_installed.append(full_code_paths[index])
 
+    else:
+        for index, install_cmd in enumerate(install_cmds):
+            proc = subprocess.Popen(install_cmd, cwd=full_code_paths[index], stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+            print('run npm install in {}'.format(full_code_paths[index]))
+            time.sleep(0.1)
+            out, err = proc.communicate()
+            if proc.returncode:
+                print("in dir:{}, executing:{}".format(full_code_paths[index], ' '.join(install_cmd[index])))
+                return False, err.decode()
+            args.success_installed.append(full_code_paths[index])
     return True, None
 
 
@@ -434,11 +447,14 @@ def main():
     install_config = config_info.get(host_platform).get(host_cpu).get('install')
     retry_times = 0
     max_retry_times = 2
+    args.success_installed = []
+    args.parallel_install = True
     while retry_times <= max_retry_times:
         result, error = _npm_install(args)
         if result:
             break
         print("npm install error, error info: %s" % error)
+        args.parallel_install = False
         if retry_times == max_retry_times:
             for error_info in error.split('\n'):
                 if error_info.endswith('debug.log'):
